add test files and lib file updates
