+++ lib/merchant_repository.rb
@@ -1,4 +1,5 @@
 require_relative 'merchant'
+require 'csv'
 
 class MerchantRepository
