+++ test/item_repository_test.rb
@@ -1,0 +1,9 @@
+require 'minitest'
+require 'minitest/autorun'
+require 'minitest/pride'
+require './lib/item_repository'
+
+class ItemRepoTest < Minitest::Test
+
+
+end
