+++ test/item_test.rb
@@ -1,0 +1,9 @@
+require 'minitest'
+require 'minitest/autorun'
+require 'minitest/pride'
+require './lib/item'
+
+class ItemTest < Minitest::Test
+
+
+end
