+++ test/merchant_repository_test.rb
@@ -1,0 +1,9 @@
+require 'minitest'
+require 'minitest/autorun'
+require 'minitest/pride'
+require './lib/merchant_repository'
+
+class MerchantRepoTest < Minitest::Test
+
+
+end
