+++ test/merchant_test.rb
@@ -1,0 +1,9 @@
+require 'minitest'
+require 'minitest/autorun'
+require 'minitest/pride'
+require './lib/merchant'
+
+class MerchantTest < Minitest::Test
+
+
+end
